+++ docs/docs/change_log.rst
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+Current (0.1.14.dev)
+
+- Added the ability for partial classes to subclass from another partial struct or ``ctypes.Structure``.
+
 0.1.13 (26/06/2025)
 -------------------